Create Auto Reload for pages assigned Rules
All we need to for the rules that are assigned to site refresh once on once loaded. This tool works great just need to add a rule that it auto reloads any page that a rule is assigned too. Is there a way to trigger F5 key or auto reload once page is half way load or starts to load then this tool runs perfect.